requirements.txt
selenium
python-dotenv
webdriver-manager
requests